+++ source/Controllers/PropertiesController.vb
@@ -31,7 +31,7 @@ Friend Class PropertiesController
 
         'view events
         AddHandler Me.View.SeriesPropertyChanged, AddressOf SeriesPropertiesViewChanged
-        AddHandler Me.View.SeriesDeleted, AddressOf SeriesDeleted
+        AddHandler Me.View.SeriesDeleted, AddressOf SeriesDeletedHandler
 
         'model events
         AddHandler _model.SeriesAdded, AddressOf UpdateView
@@ -63,8 +63,14 @@ Friend Class PropertiesController
         _model.SeriesPropertiesChangedHandler(id)
     End Sub
 
-    Private Sub SeriesDeleted(id As Integer)
-        _model.RemoveTimeSeries(id)
+    ''' <summary>
+    ''' Deletes series deleted in the view from the model
+    ''' </summary>
+    ''' <param name="ids">List of time series ids to delete</param>
+    Private Sub SeriesDeletedHandler(ids As List(Of Integer))
+        For Each id As Integer In ids
+            _model.RemoveTimeSeries(id)
+        Next
     End Sub
 
 End Class