+++ source/Views/PropertiesWindow.vb
@@ -32,10 +32,10 @@ Friend Class PropertiesWindow
     Friend Event SeriesPropertyChanged(id As Integer)
 
     ''' <summary>
-    ''' Is raised when the user deletes a row/series
+    ''' Is raised when the user deletes rows/series
     ''' </summary>
-    ''' <param name="id">Id of the time series that was deleted</param>
-    Friend Event SeriesDeleted(id As Integer)
+    ''' <param name="ids">List of ids of the time series that were deleted</param>
+    Friend Event SeriesDeleted(ids As List(Of Integer))
 
     Public Sub New()
         ' This call is required by the Windows Form Designer.
@@ -117,15 +117,23 @@ Friend Class PropertiesWindow
     End Sub
 
     ''' <summary>
-    ''' Handles the user deleting a row/series
+    ''' Handles the user deleting rows/series
     ''' </summary>
     ''' <param name="sender"></param>
     ''' <param name="e"></param>
     Private Sub DataGridView1_UserDeletingRow(sender As Object, e As DataGridViewRowCancelEventArgs) Handles DataGridView1.UserDeletingRow
-        Dim title As String = CType(e.Row.DataBoundItem, TimeSeries).Title
-        Dim id As Integer = CType(e.Row.DataBoundItem, TimeSeries).Id
-        If MsgBox($"Delete series {title}?", MsgBoxStyle.OkCancel) = MsgBoxResult.Ok Then
-            RaiseEvent SeriesDeleted(id)
+        'collect titles and ids of all selected rows
+        Dim titles As New List(Of String)
+        Dim ids As New List(Of Integer)
+        For Each row As DataGridViewRow In Me.DataGridView1.SelectedRows
+            Dim ts As TimeSeries = CType(row.DataBoundItem, TimeSeries)
+            titles.Add(ts.Title)
+            ids.Add(ts.Id)
+        Next
+        'ask for user confirmation
+        Dim result As MsgBoxResult = MsgBox($"Delete {ids.Count} series {String.Join(", ", titles)}?", MsgBoxStyle.OkCancel Or MsgBoxStyle.Exclamation)
+        If result = MsgBoxResult.Ok Then
+            RaiseEvent SeriesDeleted(ids)
         End If
         'cancel row deletion because datagridview will be refreshed from outside
         'otherwise, two rows end up being deleted